[tools] Use Babel and ES6 :elephant:
