+++ .gitignore
@@ -1,3 +1,4 @@
+lib
 lib-cov
 node_modules
 test/test-coverage.html
