+++ .jscs.json
@@ -1,6 +1,7 @@
 {
     "excludeFiles": [
         "node_modules/**",
+        "lib/**",
         "lib-cov/**"
     ],
     "requireCurlyBraces": [
