+++ package.json
@@ -1,66 +1,68 @@
 {
-    "name": "csscomb",
-    "description": "CSS coding style formatter",
-    "version": "3.1.5",
-    "homepage": "http://csscomb.com/",
-    "author": "Mikhail Troshev <mishanga@yandex-team.ru>",
-    "repository": "https://github.com/csscomb/csscomb.js",
-    "maintainers": [
-        {
-            "name": "Tony Ganch",
-            "email": "tonyganch+github@gmail.com",
-            "web": "http://tonyganch.com/"
-        }
-    ],
-    "contributors": [
-        {
-            "name": "Igor Novak",
-            "email": "bezengi@gmail.com"
-        },
-        {
-            "name": "Roman Komarov",
-            "email": "kizmarh@ya.ru"
-        },
-        {
-            "name": "Denis Payase",
-            "email": "lostsoul@yandex-team.ru"
-        },
-        {
-            "name": "Mikhail Troshev",
-            "email": "mishanga@yandex-team.ru",
-            "web": "http://mishanga.pro/"
-        },
-        {
-            "name": "Sergey Puzankov",
-            "email": "puzankov@yandex-team.ru"
-        }
-    ],
-    "engines": {
-        "node": ">= 0.10.0"
-    },
-    "dependencies": {
-        "minimist": "1.1.x",
-        "csscomb-core": "3.0.0-4",
-        "gonzales-pe": "3.0.0-29",
-        "vow": "0.4.4"
-    },
-    "devDependencies": {
-        "jshint-groups": "0.5.3",
-        "jshint": "2.3.0",
-        "jscs": "1.4.5",
-        "mocha": "1.20.1"
-    },
-    "main": "./lib/csscomb.js",
-    "files": [
-        "bin",
-        "config",
-        "lib"
-    ],
-    "bin": {
-        "csscomb": "./bin/csscomb"
-    },
-    "scripts": {
-        "test": "./node_modules/.bin/jshint-groups && ./node_modules/.bin/jscs . && node test/mocha",
-        "test-cov": "rm -rf lib-cov && jscoverage lib lib-cov && TEST_COV=true node test/mocha > ./test/test-coverage.html"
+  "name": "csscomb",
+  "description": "CSS coding style formatter",
+  "version": "3.1.5",
+  "homepage": "http://csscomb.com/",
+  "author": "Mikhail Troshev <mishanga@yandex-team.ru>",
+  "repository": "https://github.com/csscomb/csscomb.js",
+  "maintainers": [
+    {
+      "name": "Tony Ganch",
+      "email": "tonyganch+github@gmail.com",
+      "web": "http://tonyganch.com/"
     }
+  ],
+  "contributors": [
+    {
+      "name": "Igor Novak",
+      "email": "bezengi@gmail.com"
+    },
+    {
+      "name": "Roman Komarov",
+      "email": "kizmarh@ya.ru"
+    },
+    {
+      "name": "Denis Payase",
+      "email": "lostsoul@yandex-team.ru"
+    },
+    {
+      "name": "Mikhail Troshev",
+      "email": "mishanga@yandex-team.ru",
+      "web": "http://mishanga.pro/"
+    },
+    {
+      "name": "Sergey Puzankov",
+      "email": "puzankov@yandex-team.ru"
+    }
+  ],
+  "engines": {
+    "node": ">= 0.10.0"
+  },
+  "dependencies": {
+    "minimist": "1.1.x",
+    "csscomb-core": "3.0.0-4",
+    "gonzales-pe": "3.0.0-29",
+    "vow": "0.4.4"
+  },
+  "devDependencies": {
+    "babel-core": "^5.4.7",
+    "gulp-babel": "^5.1.0",
+    "jscs": "1.4.5",
+    "jshint": "2.3.0",
+    "jshint-groups": "0.5.3",
+    "mocha": "1.20.1"
+  },
+  "main": "./lib/csscomb.js",
+  "files": [
+    "bin",
+    "config",
+    "lib"
+  ],
+  "bin": {
+    "csscomb": "./bin/csscomb"
+  },
+  "scripts": {
+    "test": "./node_modules/.bin/jshint-groups && ./node_modules/.bin/jscs . && node test/mocha",
+    "test-cov": "rm -rf lib-cov && jscoverage lib lib-cov && TEST_COV=true node test/mocha > ./test/test-coverage.html"
+  }
 }
