+++ src/options/element-case.js
