+++ src/options/quotes.js
