+++ src/options/remove-empty-rulesets.js
