+++ src/options/space-after-colon.js
