+++ src/options/space-after-combinator.js
