+++ src/options/space-after-opening-brace.js
