+++ src/options/space-after-selector-delimiter.js
