+++ src/options/space-before-closing-brace.js
