+++ src/options/space-before-colon.js
