+++ src/options/space-before-combinator.js
