+++ src/options/space-before-opening-brace.js
