+++ src/options/space-before-selector-delimiter.js
