+++ src/options/strip-spaces.js
